examples/code-interpreter-agent-on-adk/requirements.txt
git+https://github.com/kubernetes-sigs/agent-sandbox.git@main#subdirectory=clients/python/agentic-sandbox-client
google-adk==1.19.0
